server/requirements.txt
fastapi==0.93.0
homeassistant-api==4.1.0
Jinja2==3.1.2
loguru==0.6.0
Pillow==8.4.0
playwright==1.31.1
pydantic>=1.6.2,<2.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pyyaml-include==1.3
PyYAML==6.0
requests==2.28.2
uvicorn[standard]==0.20.0
